Revert "Feature/mobile personal info"
