+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/model/api/ABCJobsService.kt
@@ -22,9 +22,7 @@ class ABCJobsService constructor(context: Context){
         private var LOGIN_PATH = "/login"
         private var CONFIG_PATH = "/config"
         private var ACADEMIC_INFO_PATH = "/academic-info"
-        private var PERSONAL_INFO_PATH = "/personal-info"
         private var UTILS_PATH = "/utils"
-        private var COUNTRIES_PATH = "/countries"
         private var TITLE_TYPES_PATH = "/title-types"
         private var instance: ABCJobsService? = null
 
@@ -294,79 +292,6 @@ class ABCJobsService constructor(context: Context){
         }
     }
 
-    suspend fun postPersonalInfo(token: String, personalInfo: JSONObject): Result<PersonalInfoResponse?> {
-        return try {
-            val response = suspendCoroutine<JSONObject> { cont ->
-                requestQueue.add(
-                    object : StringRequest(
-                        Method.POST, BASEURL + CANDIDATES_PATH + PERSONAL_INFO_PATH,
-                        { response -> cont.resume(JSONObject(response)) },
-                        { volleyError -> cont.resumeWithException(volleyError) }
-                    ) {
-                        override fun getBodyContentType(): String {
-                            return "application/json; charset=utf-8"
-                        }
-
-                        override fun getBody(): ByteArray {
-                            Log.d("Sending body", personalInfo.toString())
-                            return personalInfo.toString().toByteArray()
-                        }
-
-                        override fun getHeaders(): MutableMap<String, String> {
-                            val headers = HashMap<String, String>()
-                            headers["Authorization"] = "Bearer $token"
-                            return headers
-                        }
-                    }
-                )
-            }
-            if (response.getBoolean("success")) {
-                val personalInfo = deserializePersonalInfo(response)
-                Result.success(personalInfo)
-            } else {
-                val personalInfoError = deserializeAcademicInfoError(response)
-                Result.failure(personalInfoError)
-            }
-        } catch (e: Exception) {
-            Log.d("NETWORK_ERROR", e.toString())
-            Result.failure(e)
-        }
-    }
-
-    suspend fun getPersonalInfo(token: String): Result<PersonalInfoResponse?> {
-        return try {
-            val response = suspendCoroutine<JSONObject> { cont ->
-                requestQueue.add(
-                    object : StringRequest(
-                        Method.GET, BASEURL + CANDIDATES_PATH + PERSONAL_INFO_PATH,
-                        { response -> cont.resume(JSONObject(response)) },
-                        { volleyError -> cont.resumeWithException(volleyError) }
-                    ) {
-                        override fun getBodyContentType(): String {
-                            return "application/json; charset=utf-8"
-                        }
-
-                        override fun getHeaders(): MutableMap<String, String> {
-                            val headers = HashMap<String, String>()
-                            headers["Authorization"] = "Bearer $token"
-                            return headers
-                        }
-                    }
-                )
-            }
-            if (response.getBoolean("success")) {
-                val personalInfo = deserializePersonalInfo(response)
-                Result.success(personalInfo)
-            } else {
-                val personalInfoError = deserializePersonalInfoError(response)
-                Result.failure(personalInfoError)
-            }
-        } catch (e: Exception) {
-            Log.d("NETWORK_ERROR", e.toString())
-            Result.failure(e)
-        }
-    }
-
     suspend fun getTypesTitle(token: String): Result<AcademicInfoTypeResponse>{
         return try {
             val response = suspendCoroutine<JSONObject> { cont ->
@@ -401,28 +326,5 @@ class ABCJobsService constructor(context: Context){
         }
     }
 
-    suspend fun getCountries(): Result<CountriesResponse> {
-        return try {
-            val response = suspendCoroutine<JSONObject> { cont ->
-                requestQueue.add(
-                    object : StringRequest(
-                        Method.GET, BASEURL + CANDIDATES_PATH + UTILS_PATH + COUNTRIES_PATH,
-                        { response -> cont.resume(JSONObject(response)) },
-                        { volleyError -> cont.resumeWithException(volleyError) }
-                    ){}
-                )
-            }
-            if (response.getBoolean("success")) {
-                val personalInfo = deserializeCountries(response)
-                Result.success(personalInfo)
-            } else {
-                val personalInfoError = deserializeCountriesError(response)
-                Result.failure(personalInfoError)
-            }
-        } catch (e: Exception) {
-            Log.d("NETWORK_ERROR", e.toString())
-            Result.failure(e)
-        }
-    }
 
 }
