+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/model/repository/ABCJobsRepository.kt
@@ -29,22 +29,8 @@ class ABCJobsRepository(private val applicationContext: Application) {
         return ABCJobsService.getInstance(applicationContext).postConfig(token, configJson)
     }
 
-    suspend fun getPersonalInfo(token: String): Result<PersonalInfoResponse?> {
-        return ABCJobsService.getInstance(applicationContext).getPersonalInfo(token)
-    }
-
-    suspend fun postPersonalInfo(token: String, personalInfoRequest: PersonalInfoRequest): Result<PersonalInfoResponse?> {
-        val configJson = serializePersonalInfo(personalInfoRequest)
-
-        return ABCJobsService.getInstance(applicationContext).postPersonalInfo(token, configJson)
-    }
-
     suspend fun getTypeTitles(token: String): Result<AcademicInfoTypeResponse>{
         return ABCJobsService.getInstance(applicationContext).getTypesTitle(token)
-    }
-
-    suspend fun getCountries(): Result<CountriesResponse> {
-        return ABCJobsService.getInstance(applicationContext).getCountries()
     }
 
 }
