+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/view/CandidatoRegisterFragment.kt
@@ -169,6 +169,7 @@ class CandidatoRegisterFragment : Fragment(), View.OnClickListener,
             }
             true
         }
+
         return view
     }
