+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/view/LoginFragment.kt
@@ -47,7 +47,7 @@ private const val ARG_PARAM2 = "param2"
  * Use the [LoginFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class LoginFragment : Fragment(), View.OnClickListener, LoginViewModel.NavigationListener {
+class LoginFragment : Fragment(),View.OnClickListener, LoginViewModel.NavigationListener {
     private var param1: String? = null
     private var param2: String? = null
