+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/view/MainActivity.kt
@@ -18,7 +18,8 @@ import com.google.android.material.navigation.NavigationView
 import com.google.android.material.snackbar.Snackbar
 import java.util.Locale
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity()
+{
 
     private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var binding: ActivityMainBinding
@@ -33,21 +34,10 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(binding.appBarMain.toolbar)
 
         binding.appBarMain.fab.setOnClickListener { view ->
-            val navHostFragment =
-                supportFragmentManager.findFragmentById(R.id.nav_host_fragment_content_main)
-
-            when (navHostFragment?.childFragmentManager?.fragments?.firstOrNull()) {
-                is AcademicInfoFragment -> {
-                    findNavController(R.id.nav_host_fragment_content_main).navigate(R.id.academicInfoCreateFragment)
-                }
-
-                else -> {
-                    Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show()
-                }
-            }
+            Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                .setAction("Action", null).show()
         }
-
+        
         val sharedPreferences = getSharedPreferences("AppPreferences", MODE_PRIVATE)
         var selectedLanguage = Locale.getDefault().language;
         val selectedDateFormat = sharedPreferences.getString("dateFormat", "DD/MM/YYYY")
@@ -55,9 +45,9 @@ class MainActivity : AppCompatActivity() {
 
         if (selectedLanguage != null
             && selectedDateFormat != null
-            && selectedTimeFormat != null
-        ) {
-            updatePreferences(selectedLanguage, selectedDateFormat, selectedTimeFormat)
+            && selectedTimeFormat != null)
+            {
+                    updatePreferences(selectedLanguage, selectedDateFormat, selectedTimeFormat)
         }
 
 
@@ -66,22 +56,14 @@ class MainActivity : AppCompatActivity() {
         val navController = findNavController(R.id.nav_host_fragment_content_main)
         appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.nav_home, R.id.nav_personal_info, R.id.nav_academic_info, R.id.nav_slideshow
+                R.id.nav_home, R.id.nav_academic_info, R.id.nav_slideshow
             ), drawerLayout
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
 
-    fun hideButton() {
-        binding.appBarMain.fab.hide()
-    }
-
-    fun updatePreferences(
-        selectedLanguage: String,
-        selectedDateFormat: String,
-        selectedTimeFormat: String
-    ) {
+    fun updatePreferences(selectedLanguage: String, selectedDateFormat: String, selectedTimeFormat: String) {
         val sharedPreferences = getSharedPreferences("AppPreferences", MODE_PRIVATE)
         val editor = sharedPreferences.edit()
 
@@ -96,7 +78,6 @@ class MainActivity : AppCompatActivity() {
                     recreate()
                 }
             }
-
             "Español" -> {
                 if (currentLanguage != "es") {
                     setLocale("es")
