--- mobile/app/src/main/java/com/example/abc_jobs_alpaca/viewmodel/AcademicInfoCreateViewModel.kt
@@ -1,8 +1,0 @@
-package com.example.abc_jobs_alpaca.viewmodel
-
-import androidx.lifecycle.ViewModel
-import com.example.abc_jobs_alpaca.model.repository.ABCJobsRepository
-
-class AcademicInfoCreateViewModel( private val abcJobsRepository: ABCJobsRepository): ViewModel() {
-
-}
