+++ mobile/app/src/main/java/com/example/abc_jobs_alpaca/viewmodel/AcademicInfoViewModel.kt
@@ -3,9 +3,9 @@ package com.example.abc_jobs_alpaca.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.abc_jobs_alpaca.model.repository.ABCJobsRepository
 
-class AcademicInfoViewModel (private val abcJobsRepository: ABCJobsRepository) : ViewModel(){
+class AcademicInfoViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
     private val _text = MutableLiveData<String>().apply {
         value = "This is academic info Fragment"
     }
