--- mobile/app/src/main/res/drawable/black_rounded_layout_background.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<inset xmlns:android="http://schemas.android.com/apk/res/android"
-    android:insetTop="0dp">
-
-<selector>
-<item>
-    <shape>
-        <solid android:color="@android:color/transparent" />
-
-        <stroke
-            android:color="@color/black_half_alpha"
-            android:width="1dp"
-            />
-        <corners android:radius="48dp" />
-    </shape>
-</item>
-</selector>
-    </inset>
