+++ mobile/app/src/main/res/navigation/mobile_navigation.xml
@@ -26,16 +26,9 @@
         android:name="com.example.abc_jobs_alpaca.view.PreferencesFragment"
         android:label="@string/preferencias_del_sistema"
         tools:layout="@layout/fragment_preferences"/>
-
     <fragment
         android:id="@+id/academicInfoCreateFragment"
-        android:name="com.example.abc_jobs_alpaca.view.AcademicInfoCreateFragment"
+        android:name="com.example.abc_jobs_alpaca.AcademicInfoCreateFragment"
         android:label="fragment_academic_info_create"
         tools:layout="@layout/fragment_academic_info_create" />
-
-    <fragment
-        android:id="@+id/nav_personal_info"
-        android:name="com.example.abc_jobs_alpaca.view.PersonalInfoFragment"
-        android:label="@string/menu_academic_info"
-        tools:layout="@layout/fragment_personal_info" />
 </navigation>
