+++ mobile/app/src/main/res/values-en/strings.xml
@@ -58,23 +58,9 @@
     <string name="successful_toast_message">Account created successfully</string>
     <string name="successful_login_toast_message">Login successful</string>
     <string name="toast_message_login_failed">Login failed, Email not registered</string>
-    <string name="cancel_preferences_button">Cancel</string>
-    <string name="menu_academic_info">Academic information</string>
-    <string name="menu_personal_info">Personal information</string>
-    <string name="preferencias_del_sistema">System Preferences</string>
-    <string name="no_personal_info">User does not have personal information yet, you can add it using the button located in the bottom right corner.</string>
-    <string name="full_name_header">Full Name</string>
-    <string name="full_name_placeholder">Name and Last Name</string>
-    <string name="city_header">City of living</string>
-    <string name="city_placeholder">Write city</string>
-    <string name="address_header">Residence Address</string>
-    <string name="address_placeholder">Write address</string>
-    <string name="phone_header">Phone number</string>
-    <string name="phone_placeholder">Write phone number</string>
-    <string name="bio_header">Bio</string>
-    <string name="bio_placeholder">Write bio</string>
-    <string name="save_personal_info">Save</string>
-    <string name="birthday_header">Birth date</string>
-    <string name="birthday_placeholder">01/01/1990</string>
-    <string name="country_header">Country of residence</string>
+    <string name="cancel_preferences_button">Cancelar</string>
+    <string name="menu_academic_info">Información académica</string>
+    <string name="preferencias_del_sistema">Preferencias del sistema</string>
+
+
 </resources>
