+++ mobile/app/src/main/res/values/colors.xml
@@ -3,7 +3,6 @@
     <color name="accent">#5885DD</color>
     <color name="gray">#D9D9D9</color>
     <color name="black">#FF000000</color>
-    <color name="black_half_alpha">#80000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="gray_300">#F2F2F2</color>
     <color name="black_overlay">#66000000</color>
